requirements.txt
numpy>=1.26.0
requests>=2.31.0
pandas>=2.0.3
google-generativeai>=0.3.2
fastapi>=0.111.0
uvicorn>=0.29.0 